Remove unnecesary module.
Keep all deserialization logic together, it's easier to manage.
